Make changes to improve PR quality
